Remove SQLite compose config, default to a postgres container for dev
This is more prod-like. The application does still support an SQLite
database with the correct envvar set. The compose config should
encourage a default that's prod-like.

TODO: Think about how to enable an easy switch to a compose config that
supports sqlite? Profiles? Can you assign a service to _NOT_ be in a
profile, e.g. `profiles: ["!sqlite"]`? I don't think so... other
options??